chore: refactor processing logic
Move the processing logic from the cli into vectorizer so we can
expose it in the library.
