+++ projects/pgai/pgai/vectorizer/__init__.py
@@ -1,8 +1,10 @@
 from .create_vectorizer import CreateVectorizer
+from .processor import Processor
 from .vectorizer import Vectorizer, Worker
 
 __all__ = [
     "Vectorizer",
     "Worker",
     "CreateVectorizer",
+    "Processor",
 ]
